pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/f3"]

[project]
name = "f3"
version = "1.0.0"
description = "Fast Feature Fields (F3): A Predictive Representation of Events"
authors = [
    {name = "Richeek Das"},
    {name = "Pratik Chaudhari"},
    {name = "Kostas Daniilidis"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate==1.10.1",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "attrs==25.3.0",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "click==8.3.0",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "datasets==4.1.1",
    "dill==0.4.0",
    "evaluate==0.4.6",
    "filelock==3.13.1",
    "fonttools==4.60.0",
    "frozenlist==1.7.0",
    "fsspec==2024.6.1",
    "gitdb==4.0.12",
    "GitPython==3.1.45",
    "h5py==3.14.0",
    "hdf5plugin==5.1.0",
    "hf-xet==1.1.10",
    "huggingface-hub==0.35.1",
    "idna==3.10",
    "imageio==2.37.0",
    "Jinja2==3.1.4",
    "joblib==1.5.2",
    "kiwisolver==1.4.9",
    "MarkupSafe==2.1.5",
    "matplotlib==3.10.6",
    "mpmath==1.3.0",
    "multidict==6.6.4",
    "multiprocess==0.70.16",
    "networkx==3.3",
    "numpy==2.1.2",
    "nvidia-cublas-cu12>=12.6.4.1",
    "nvidia-cuda-cupti-cu12>=12.6.80",
    "nvidia-cuda-nvrtc-cu12>=12.6.77",
    "nvidia-cuda-runtime-cu12>=12.6.77",
    "nvidia-cudnn-cu12>=9.10.2.21",
    "nvidia-cufft-cu12>=11.3.0.4",
    "nvidia-cufile-cu12>=1.11.1.6",
    "nvidia-curand-cu12>=10.3.7.77",
    "nvidia-cusolver-cu12>=11.7.1.2",
    "nvidia-cusparse-cu12>=12.5.4.2",
    "nvidia-cusparselt-cu12>=0.7.1",
    "nvidia-nccl-cu12>=2.27.3",
    "nvidia-nvjitlink-cu12>=12.6.85",
    "nvidia-nvtx-cu12>=12.6.77",
    "opencv-python==4.12.0.88",
    "packaging==25.0",
    "pandas==2.3.2",
    "pillow==11.0.0",
    "platformdirs==4.4.0",
    "propcache==0.3.2",
    "protobuf==6.32.1",
    "psutil==7.1.0",
    "pyarrow==21.0.0",
    "pydantic==2.11.9",
    "pydantic_core==2.33.2",
    "pyparsing==3.2.5",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "PyYAML==6.0.2",
    "regex==2025.9.18",
    "requests==2.32.5",
    "safetensors==0.6.2",
    "scikit-learn==1.7.2",
    "scipy==1.16.2",
    "sentry-sdk==2.38.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "sympy==1.13.3",
    "threadpoolctl==3.6.0",
    "timm==1.0.20",
    "tokenizers==0.22.1",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "tqdm==4.67.1",
    "transformers==4.56.2",
    "triton==3.4.0",
    "typing-inspection==0.4.1",
    "typing_extensions==4.12.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "wandb==0.22.0",
    "xformers==0.0.32.post2",
    "xxhash==3.5.0",
    "yarl==1.20.1",
]
